[REFACTOR] Extract and centralize response/error/idempotency helpers
